template1/project_name/requirements/development.txt
--requirement=base.txt

Fabric==1.4.1
coverage==3.5.1
django-coverage==1.2.1
django-debug-toolbar==0.9.1
pep8==0.6.1
django-appconf==0.5
pycrypto==2.5
ssh==1.7.13
